Added enhancer, reducer and actions Still not working, the consistency between side bar notifications value and in the notificarions, still needs to work
