+++ src/options/notifications/actions.js
@@ -1,0 +1,5 @@
+import { createAction } from 'redux-act'
+
+export const unreadMessagesCount = createAction(
+    'notifications/unreadMessagesCount',
+)
