+++ src/options/notifications/selectors.js
@@ -1,0 +1,8 @@
+import { createSelector } from 'reselect'
+
+export const notifications = state => state.notifications
+
+export const unreadMessagesCount = createSelector(
+    notifications,
+    state => state.unreadMessagesCount,
+)
