+++ src/options/privacy/reducer.js
@@ -1,5 +1,4 @@
 import { createReducer } from 'redux-act'
-
 import * as actions from './actions'
 
 const defState = {
