+++ src/options/store.js
@@ -4,11 +4,13 @@ import thunk from 'redux-thunk'
 import * as imports from './imports'
 import * as blacklist from './blacklist'
 import * as privacy from './privacy'
+import * as notifications from './notifications'
 
 const rootReducer = combineReducers({
     blacklist: blacklist.reducer,
     imports: imports.reducer,
     privacy: privacy.reducer,
+    notifications: notifications.reducer,
 })
 
 export default function configureStore({ ReduxDevTools = undefined } = {}) {
@@ -16,6 +18,7 @@ export default function configureStore({ ReduxDevTools = undefined } = {}) {
         blacklist.enhancer,
         imports.enhancer,
         privacy.enhancer,
+        notifications.enhancer,
         applyMiddleware(thunk),
     ]
